backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
google-cloud-storage==2.16.0
google-cloud-spanner==3.42.0
google-cloud-aiplatform==1.49.0
google-cloud-functions==1.12.0
google-api-python-client==2.124.0
pdfplumber==0.10.3
python-docx==1.1.0
spacy
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0.tar.gz
sentence-transformers==2.7.0
openai==1.31.0
google-generativeai==0.5.2
faiss-cpu==1.8.0
chromadb==0.4.24
gunicorn==22.0.0
python-multipart==0.0.9
httpx==0.27.0
pandas==2.2.2
sqlalchemy==2.0.30
black==24.4.2
isort==5.13.2
mypy==1.10.0
types-requests==2.31.0.20240406
python-dotenv==1.0.1 